+++ preset/eslintrc/typescript.js
@@ -1,1 +1,18 @@
-// TODO
+module.exports = {
+  root: true,
+  env: {
+    browser: true,
+    node: true,
+    es6: true,
+  },
+  parserOptions: {
+    parser: '@typescript-eslint/parser',
+    sourceType: 'module'
+  },
+  plugins: [
+    '@typescript-eslint'
+  ],
+  extends: [
+    'plugin:@typescript-eslint/recommended'
+  ]
+}