+++ preset/jest/typescript.js
@@ -1,7 +1,7 @@
-
 module.exports = {
   moduleFileExtensions: [
     "js",
+    "ts",
     "json",
   ],
   transform: {